R6 Object for Feature Extraction.

Project calculates features from longitudinal data for each grouping variable individually.

Format

R6Class object.

Usage

my_project = Project$new("my_project")

Arguments

For Project$new():

project_name:

('character(1)'): A user defined name of the project. All necessary data will be saved on the path: ./projects/project_name/

load:

(`logical(1)`): If TRUE, an existing project will be loaded.

Details

All datasets and feature functions are saved in this R6 object. Project heavily relies on the R-package batchtools. Data will be saved as single RDS files (for each grouping variable) and feature functions are calculated on each single dataset. A big advantage of this method is that it scales nicely for larger datasets. Data is only loaded into RAM, when needed.

Fields

project_name:

(`character(1)`): The projects name.

dir:

(`character(1)`): The projects directory.

group_by:

(`character(1)`): The column on which to group by.

reg:

(`registry`): batchtools registry.

Methods

add_data(dataframe, group_by)

[dataframe: (`data.frame`)] A dataframe which shall be added to the R6 object.
[group_by: (`character(1)`)] The grouping variable of the dataframe.

This method writes single RDS files (this can be parallelized with foreach) and adds them as batchtools problems for each grouping variable of a dataframe. After that, batchtools experiments will be added too.

preprocess_data(fun)

[fun: (`function`)] A function, which has a dataframe as input and a dataframe as output.

This method loads the RDS files and applies this function on them. The old RDS files are overwritten and the batchtools problems and experiments are updated.

remove_data(data)

[data: (`character(1)`)] The grouping variable's name.

This method deletes the RDS file and batchtools problem of one single grouping variable.

add_feature(fun)

[fun: (`function`)] A function, which has a dataframe as input and a named vector as output.

This method adds a function as batchtools algorithm. After that, batchtools experiments are added too.

remove_feature(fun)

[fun: (`function | character(1)`)] A function (or the name of the function as character) which shall be removed.

This method removes the batchtools algorithms and experiments corresponding to the given function.

calc_features()

This method calculates all features on all datasets. Internally, it submits all batchtools jobs, which are not done yet.

get_project_status()

This method gives an overview over which features are done on which datasets.

collect_results()

This method returns a dataframe with the calculated features.

plot()

[internal] method to print the R6 object.

clone()

[internal] method to clone the R6 object.

initialize()

[internal] method to initialize the R6 object.

Examples

unlink("projects/my_project", recursive = TRUE)
my_project = Project$new("my_project")

#> No readable configuration file found

#> Created registry in 'L:/Users/auj/Desktop/Git/fxtract/docs/reference/projects/my_project/reg' using cluster functions 'Interactive'

my_project$add_data(iris, group_by = "Species")

#> Warning: executing %dopar% sequentially: no parallel backend registered

#> 
#> Attache Paket: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

#> Saving raw RDS file setosa.RDS on disk.

#> Saving raw RDS file versicolor.RDS on disk.

#> Saving raw RDS file virginica.RDS on disk.

#> Adding problem 'setosa'

#> Adding problem 'versicolor'

#> Adding problem 'virginica'

fun = function(data) {
  c(mean_sepal_length = mean(data$Sepal.Length))
}
my_project$add_feature(fun)

#> Exporting new objects: 'fun' ...

#> Adding algorithm 'fun'

#> Adding 1 experiments ('setosa'[1] x 'fun'[1] x repls[1]) ...

#> Adding 1 experiments ('versicolor'[1] x 'fun'[1] x repls[1]) ...

#> Adding 1 experiments ('virginica'[1] x 'fun'[1] x repls[1]) ...

my_project$calc_features()

#> Submitting 3 jobs in 3 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Generating problem instance for problem 'setosa' ...

#> ### [bt]: Applying algorithm 'fun' on problem 'setosa' for job 1 (seed = 2647) ...

#> ### [bt]: Generating problem instance for problem 'versicolor' ...

#> ### [bt]: Applying algorithm 'fun' on problem 'versicolor' for job 2 (seed = 2648) ...

#> ### [bt]: Generating problem instance for problem 'virginica' ...

#> ### [bt]: Applying algorithm 'fun' on problem 'virginica' for job 3 (seed = 2649) ...

my_project$collect_results()

#>      Species mean_sepal_length
#> 1     setosa             5.006
#> 2 versicolor             5.936
#> 3  virginica             6.588

unlink("projects/my_project", recursive = TRUE)