Calculates a given function grouped by group_col. For each group variable one single value is returned.

Calculates a given function grouped by group_col. For each group variable one single value is returned.

calcFeatureByGroup(data, group_col, fun, colname, check_fun = TRUE)

Arguments

data
dataframe.

group_col
character. Name of column, which contains identifiers on which the dataset should be grouped by. E.g. different user IDs.

fun
function. Must be a function, which has a dataframe as input and a vector of length 1 as output.

colname
character. Column name of the new column.

check_fun
logical. If TRUE, fun(data) will be evaluated and checked, if the outcome is of correct form. Set to FALSE, if evaluation on the whole dataset takes too long.

Value

dataframe with columns group_col and colname.

Examples

fun = function(data) {
  nrow(dplyr::filter(data, RUNNING_TASKS_baseActivity_mPackage == "com.android.chrome"))
}
calcFeatureByGroup(data = studentlife.small, group_col = "userId", fun = fun,
  colname = "number_uses_chrome")

#> # A tibble: 3 x 2
#> # Groups:   userId [3]
#>   userId number_uses_chrome
#>   <chr>               <int>
#> 1 1                     199
#> 2 2                       9
#> 3 3                     198