Calculates a given function on the dataset for each studyDay and grouped by the column 'userId'. Needs a function, which has a dataframe as input and a singular value as output. A summary function, which summarizes the results for each day, will be evaluated for each 'userId'.

Calculates a given function on the dataset for each studyDay and grouped by the column 'userId'. Needs a function, which has a dataframe as input and a singular value as output. A summary function, which summarizes the results for each day, will be evaluated for each 'userId'.

calcFeaturePerStudyDayAndUserId(data, fun, colname, summary_fun, max_study_day,
  impute_empty_day = NA, export_results_per_day = FALSE)

Arguments

data
dataframe. Dataframe containing column named 'userId', which should be the unique identifier for each ID.

fun
function. Must be a function, which has a dataframe as input and a vector (of length nrow(dataframe)) as output.

colname
character. Column name of the new column.

summary_fun
function. Must be a function, which has a vector as input and a singular value as output.

max_study_day
integer. Function will be evaluated until the max study day.

impute_empty_day
can be character, factor, or numerical. Results on days, which did not have any data, will be imputed with this value.

export_results_per_day
logical. If TRUE, an additional dataframe, where fun is evaluated at each studyDay, is returned.

Value

dataframe with columns userId and colname. One value for each userId.

See also

Examples

fun = function(data) {
  nrow(dplyr::filter(data, RUNNING_TASKS_baseActivity_mPackage == "com.android.chrome"))
} #number uses google chrome
summary_fun = function(x) mean(x, na.rm = TRUE)
data = addStudyDay(studentlife.small)

#> If you have data with different users, be aware, that this function ignores that! See addStudyDayPerUserId().

calcFeaturePerStudyDayAndUserId(data = data, fun = fun, colname = "mean_uses_chrome_per_day",
 summary_fun = summary_fun)

#> max_study_day will be automatically calculated from the given data.

#>   userId mean_uses_chrome_per_day
#> 1      1                 33.16667
#> 2      2                  1.50000
#> 3      3                 33.00000