Adds a variable (or a set of variables) to a dataset by a sliding window approach.

Adds a variable (or a set of variables) to a dataset by a sliding window approach.

slidingWindow(data, utc_col = character(1), fun, steps, time_in_sec,
  check_fun = TRUE, eval_at_rows = numeric(0), unit = "s")

Arguments

data
dataframe.

utc_col
character. Column name of the UTC timestamp.

fun
function. A function that has `data`` as input and a dataframe with 1 row as output.

steps
integer. Number of steps which the sliding window should go back. Use this for data with fixed sample rate.

time_in_sec
integer. Number of seconds which the sliding window should go back. Use this for data with variable sample rate.

check_fun
logical. If TRUE, fun(data) will be evaluated and checked, if the outcome is of correct form. Set to FALSE, if evaluation on the whole dataset takes too long.

eval_at_rows
integer. Integer vector at which rows the function should be executed. Defaults to `1:nrow(data)`.

unit
unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).

Value

dataframe with added column(s).

See also

Examples

fun = function(x) data.frame(mean.accuracy.last30 = mean(x$accuracy, na.rm = TRUE),
  max.accuracy.last30 = max(x$accuracy, na.rm = TRUE))
data = studentlife.small[1:30, ] #needs date_time variable
slidingWindow(data, fun = fun, time_in_sec = 60 * 60,
  utc_col = "timestamp") #mean accuracy of last hour

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |                                                                      |   0%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=====                                                                 |   7%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=======                                                               |  10%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=========                                                             |  13%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |============                                                          |  17%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |==============                                                        |  20%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |================                                                      |  23%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |===================                                                   |  27%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=====================                                                 |  30%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=======================                                               |  33%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |==========================                                            |  37%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |============================                                          |  40%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |==============================                                        |  43%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=================================                                     |  47%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |===================================                                   |  50%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=====================================                                 |  53%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |========================================                              |  57%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |==========================================                            |  60%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |============================================                          |  63%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |===============================================                       |  67%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=================================================                     |  70%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |===================================================                   |  73%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |======================================================                |  77%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |========================================================              |  80%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |==========================================================            |  83%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=============================================================         |  87%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |===============================================================       |  90%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |=================================================================     |  93%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |====================================================================  |  97%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |======================================================================| 100%

#>    provider network_type accuracy latitude longitude altitude bearing speed
#> 1      <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 2      <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 3      <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 4      <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 5      <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 6      <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 7      <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 8      <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 9      <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 10     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 11     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 12     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 13     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 14     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 15     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 16     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 17     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 18     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 19     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 20     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 21     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 22     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 23     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 24     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 25     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 26     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 27     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 28     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 29     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#> 30     <NA>         <NA>       NA       NA        NA       NA      NA    NA
#>    travelstate userId  timestamp                               device
#> 1         <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 2         <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 3         <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 4         <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 5         <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 6         <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 7         <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 8         <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 9         <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 10        <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 11        <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 12        <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 13        <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 14        <NA>     00 1364095883 1977b545-a88f-4903-a7ae-2c434de4be49
#> 15        <NA>     00 1364097083 1977b545-a88f-4903-a7ae-2c434de4be49
#> 16        <NA>     00 1364097083 1977b545-a88f-4903-a7ae-2c434de4be49
#> 17        <NA>     00 1364097083 1977b545-a88f-4903-a7ae-2c434de4be49
#> 18        <NA>     00 1364097083 1977b545-a88f-4903-a7ae-2c434de4be49
#> 19        <NA>     00 1364097083 1977b545-a88f-4903-a7ae-2c434de4be49
#> 20        <NA>     00 1364098283 1977b545-a88f-4903-a7ae-2c434de4be49
#> 21        <NA>     00 1364098283 1977b545-a88f-4903-a7ae-2c434de4be49
#> 22        <NA>     00 1364098283 1977b545-a88f-4903-a7ae-2c434de4be49
#> 23        <NA>     00 1364098283 1977b545-a88f-4903-a7ae-2c434de4be49
#> 24        <NA>     00 1364098283 1977b545-a88f-4903-a7ae-2c434de4be49
#> 25        <NA>     00 1364099483 1977b545-a88f-4903-a7ae-2c434de4be49
#> 26        <NA>     00 1364099483 1977b545-a88f-4903-a7ae-2c434de4be49
#> 27        <NA>     00 1364099483 1977b545-a88f-4903-a7ae-2c434de4be49
#> 28        <NA>     00 1364099483 1977b545-a88f-4903-a7ae-2c434de4be49
#> 29        <NA>     00 1364099483 1977b545-a88f-4903-a7ae-2c434de4be49
#> 30        <NA>     00 1364100683 1977b545-a88f-4903-a7ae-2c434de4be49
#>                          RUNNING_TASKS_baseActivity_mClass
#> 1                          org.bewellapp.dashBoardActivity
#> 2          com.android.packageinstaller.InstallAppProgress
#> 3                      com.google.android.apps.chrome.Main
#> 4                      com.google.android.gm.GmailActivity
#> 5                           com.android.launcher2.Launcher
#> 6              com.android.systemui.recent.RecentsActivity
#> 7                        com.android.gallery3d.app.Gallery
#> 8   com.google.android.apps.paco.ExperimentManagerActivity
#> 9                            com.android.settings.Settings
#> 10                         com.android.deskclock.DeskClock
#> 11 com.android.settings.Settings$PowerUsageSummaryActivity
#> 12       com.android.contacts.activities.DialtactsActivity
#> 13                    edu.dartmouth.cs.camera.MainActivity
#> 14                          com.android.phone.InCallScreen
#> 15                       com.example.android.apis.ApiDemos
#> 16                          com.android.launcher2.Launcher
#> 17     com.google.android.gm.ConversationListActivityGmail
#> 18 com.android.settings.Settings$PowerUsageSummaryActivity
#> 19                          com.android.phone.InCallScreen
#> 20                       com.example.android.apis.ApiDemos
#> 21                          com.android.launcher2.Launcher
#> 22     com.google.android.gm.ConversationListActivityGmail
#> 23 com.android.settings.Settings$PowerUsageSummaryActivity
#> 24                          com.android.phone.InCallScreen
#> 25                       com.example.android.apis.ApiDemos
#> 26                          com.android.launcher2.Launcher
#> 27     com.google.android.gm.ConversationListActivityGmail
#> 28 com.android.settings.Settings$PowerUsageSummaryActivity
#> 29                          com.android.phone.InCallScreen
#> 30     com.google.android.gm.ConversationListActivityGmail
#>    RUNNING_TASKS_baseActivity_mPackage RUNNING_TASKS_id
#> 1                        org.bewellapp              249
#> 2         com.android.packageinstaller              248
#> 3                   com.android.chrome              245
#> 4                com.google.android.gm              244
#> 5                 com.android.launcher                2
#> 6                 com.android.systemui              240
#> 7         com.google.android.gallery3d              241
#> 8         com.google.android.apps.paco              247
#> 9                 com.android.settings              243
#> 10        com.google.android.deskclock              242
#> 11                com.android.settings               76
#> 12                com.android.contacts              235
#> 13             edu.dartmouth.cs.camera              232
#> 14                   com.android.phone               61
#> 15            com.example.android.apis              259
#> 16                com.android.launcher                2
#> 17               com.google.android.gm              256
#> 18                com.android.settings               76
#> 19                   com.android.phone               61
#> 20            com.example.android.apis              259
#> 21                com.android.launcher                2
#> 22               com.google.android.gm              256
#> 23                com.android.settings               76
#> 24                   com.android.phone               61
#> 25            com.example.android.apis              259
#> 26                com.android.launcher                2
#> 27               com.google.android.gm              256
#> 28                com.android.settings               76
#> 29                   com.android.phone               61
#> 30               com.google.android.gm              256
#>    RUNNING_TASKS_numActivities RUNNING_TASKS_numRunning
#> 1                            1                        1
#> 2                            1                        1
#> 3                            1                        1
#> 4                            1                        1
#> 5                            1                        1
#> 6                            1                        1
#> 7                            3                        3
#> 8                            4                        4
#> 9                            2                        2
#> 10                           1                        1
#> 11                           1                        1
#> 12                           1                        1
#> 13                           2                        2
#> 14                           1                        1
#> 15                           3                        3
#> 16                           1                        1
#> 17                           1                        1
#> 18                           1                        0
#> 19                           1                        1
#> 20                           3                        3
#> 21                           1                        1
#> 22                           1                        1
#> 23                           1                        0
#> 24                           1                        1
#> 25                           3                        3
#> 26                           1                        1
#> 27                           1                        1
#> 28                           1                        0
#> 29                           1                        1
#> 30                           1                        1
#>                           RUNNING_TASKS_topActivity_mClass
#> 1                          org.bewellapp.dashBoardActivity
#> 2          com.android.packageinstaller.InstallAppProgress
#> 3                      com.google.android.apps.chrome.Main
#> 4                      com.google.android.gm.GmailActivity
#> 5                           com.android.launcher2.Launcher
#> 6              com.android.systemui.recent.RecentsActivity
#> 7                        com.android.gallery3d.app.Gallery
#> 8     com.google.android.apps.paco.FindExperimentsActivity
#> 9                         com.android.settings.SubSettings
#> 10                         com.android.deskclock.DeskClock
#> 11 com.android.settings.Settings$PowerUsageSummaryActivity
#> 12       com.android.contacts.activities.DialtactsActivity
#> 13                 edu.dartmouth.cs.camera.ProfileActivity
#> 14                          com.android.phone.InCallScreen
#> 15                       com.example.android.apis.ApiDemos
#> 16                          com.android.launcher2.Launcher
#> 17     com.google.android.gm.ConversationListActivityGmail
#> 18 com.android.settings.Settings$PowerUsageSummaryActivity
#> 19                          com.android.phone.InCallScreen
#> 20                       com.example.android.apis.ApiDemos
#> 21                          com.android.launcher2.Launcher
#> 22     com.google.android.gm.ConversationListActivityGmail
#> 23 com.android.settings.Settings$PowerUsageSummaryActivity
#> 24                          com.android.phone.InCallScreen
#> 25                       com.example.android.apis.ApiDemos
#> 26                          com.android.launcher2.Launcher
#> 27     com.google.android.gm.ConversationListActivityGmail
#> 28 com.android.settings.Settings$PowerUsageSummaryActivity
#> 29                          com.android.phone.InCallScreen
#> 30     com.google.android.gm.ConversationListActivityGmail
#>    RUNNING_TASKS_topActivity_mPackage mean.accuracy.last30 max.accuracy.last30
#> 1                       org.bewellapp                   NA                  NA
#> 2        com.android.packageinstaller                  NaN                -Inf
#> 3                  com.android.chrome                  NaN                -Inf
#> 4               com.google.android.gm                  NaN                -Inf
#> 5                com.android.launcher                  NaN                -Inf
#> 6                com.android.systemui                  NaN                -Inf
#> 7        com.google.android.gallery3d                  NaN                -Inf
#> 8        com.google.android.apps.paco                  NaN                -Inf
#> 9                com.android.settings                  NaN                -Inf
#> 10       com.google.android.deskclock                  NaN                -Inf
#> 11               com.android.settings                  NaN                -Inf
#> 12               com.android.contacts                  NaN                -Inf
#> 13            edu.dartmouth.cs.camera                  NaN                -Inf
#> 14                  com.android.phone                  NaN                -Inf
#> 15           com.example.android.apis                  NaN                -Inf
#> 16               com.android.launcher                  NaN                -Inf
#> 17              com.google.android.gm                  NaN                -Inf
#> 18               com.android.settings                  NaN                -Inf
#> 19                  com.android.phone                  NaN                -Inf
#> 20           com.example.android.apis                  NaN                -Inf
#> 21               com.android.launcher                  NaN                -Inf
#> 22              com.google.android.gm                  NaN                -Inf
#> 23               com.android.settings                  NaN                -Inf
#> 24                  com.android.phone                  NaN                -Inf
#> 25           com.example.android.apis                  NaN                -Inf
#> 26               com.android.launcher                  NaN                -Inf
#> 27              com.google.android.gm                  NaN                -Inf
#> 28               com.android.settings                  NaN                -Inf
#> 29                  com.android.phone                  NaN                -Inf
#> 30              com.google.android.gm                  NaN                -Inf