R6 Object for Feature Extraction.

Xtractor calculates features from raw data for each ID of a grouping variable individually with batchtools.

Format

R6Class object.

Usage

xtractor = Xtractor$new("xtractor")

Arguments

For Xtractor$new():

name:

('character(1)'): A user defined name of the Xtractor. All necessary data will be saved on the path: ./fxtract_files/name/

load:

(`logical(1)`): If TRUE, an existing Xtractor will be loaded.

Details

All datasets and feature functions are saved in this R6 object. Xtractor heavily relies on the R-package batchtools. Data will be saved as single RDS files (for each ID) and feature functions are calculated on each single dataset. A big advantage of this method is that it scales nicely for larger datasets. Data is only read into RAM, when needed.

Fields

name:

(`character(1)`): The Xtractors name.

dir:

(`character(1)`): The directory where files are saved.

group_by:

(`character(1)`): The column on which to group by.

reg:

(`registry`): batchtools registry.

perc_done:

(`numeric(1)`): Active binding. Percentage of finished calculations.

error_messages:

(`data.frame()`): Active binding. A dataframe with information about error messages.

log_files:

(`list()`): Active binding. A list with the log files which were created due to errors.

datasets:

(`character()`): Active binding. A character vector with the IDs of the grouping variable.

features:

(`character()`): Active binding. A character vector with the feature functions which were added.

status:

(`data.frame()`): Active binding. A dataframe with an overview over which features are done on which datasets.

results:

(`data.frame()`): Active binding. A dataframe with all calculated features of all IDs.

Methods

add_data(dataframe, group_by)

[dataframe: (`data.frame`)] A dataframe which shall be added to the R6 object.
[group_by: (`character(1)`)] The grouping variable's name of the dataframe.
This method writes single RDS files (this can be parallelized with foreach) and adds them as batchtools problems for each ID of the grouping variable. After that, batchtools experiments will be added too.

preprocess_data(fun)

[fun: (`function`)] A function, which has a dataframe as input and a dataframe as output.
This method loads the RDS files and applies this function on them. The old RDS files are overwritten and the batchtools problems and experiments are updated.

remove_data(data)

[data: (`character(1)`)] An ID of the grouping variable.
This method deletes the RDS file and batchtools problem of one single ID of the grouping variable.

get_data(data)

[data: (`character()`)] One ore many IDs of the grouping variable.
This method returns one dataframe with the chosen IDs.

add_feature(fun)

[fun: (`function`)] A function, which has a dataframe as input and a named vector as output.
This method adds a function as batchtools algorithm. After that, batchtools experiments are added too.

remove_feature(fun)

[fun: (`function | character(1)`)] A function (or the name of the function as character) which shall be removed.
This method removes the batchtools algorithms and experiments corresponding to the given function.

get_feature(fun)

[fun: (`character(1)`)] The name of a function as character.
This method reads the RDS file of the function. Useful for debugging after loading an Xtractor.

calc_features()

This method calculates all features on all datasets. Internally, it submits all batchtools jobs, which are not done yet.

plot()

[internal] method to print the R6 object.

clone()

[internal] method to clone the R6 object.

initialize()

[internal] method to initialize the R6 object.

Examples

# NOT RUN {
  xtractor = Xtractor$new("xtractor")
  xtractor$add_data(iris, group_by = "Species")
  xtractor$datasets
  fun = function(data) {
    c(mean_sepal_length = mean(data$Sepal.Length))
  }
  xtractor$add_feature(fun)
  xtractor$features
  xtractor$calc_features()
  xtractor$results
  xtractor$perc_done
  unlink("fxtract_files/xtractor", recursive = TRUE)
# }