Adds a variable (or a set of variables) to a dataset by a sliding window approach.

Adds a variable (or a set of variables) to a dataset by a sliding window approach.

slidingWindow(data, utc_col = character(1), fun, steps, time_in_sec,
  check_fun = TRUE, eval_at_rows = numeric(0), unit = "s")

Arguments

data
dataframe.

utc_col
character. Column name of the UTC timestamp.

fun
function. A function that has `data`` as input and a dataframe with 1 row as output.

steps
integer. Number of steps which the sliding window should go back. Use this for data with fixed sample rate.

time_in_sec
integer. Number of seconds which the sliding window should go back. Use this for data with variable sample rate.

check_fun
logical. If TRUE, fun(data) will be evaluated and checked, if the outcome is of correct form. Set to FALSE, if evaluation on the whole dataset takes too long.

eval_at_rows
integer. Integer vector at which rows the function should be executed. Defaults to `1:nrow(data)`.

unit
unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).

Value

dataframe with added column(s).

See also

Examples

fun = function(x) data.frame(mean.accuracy.last30 = mean(x$accuracy, na.rm = TRUE),
  max.accuracy.last30 = max(x$accuracy, na.rm = TRUE))
data = studentlife.small[1:30, ] #needs date_time variable
slidingWindow(data, fun = fun, time_in_sec = 60 * 60,
  utc_col = "timestamp") #mean accuracy of last hour

#>   |                                                                              |                                                                      |   0%  |                                                                              |=====                                                                 |   7%  |                                                                              |=======                                                               |  10%  |                                                                              |=========                                                             |  13%  |                                                                              |============                                                          |  17%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  23%  |                                                                              |===================                                                   |  27%  |                                                                              |=====================                                                 |  30%  |                                                                              |=======================                                               |  33%  |                                                                              |==========================                                            |  37%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  43%  |                                                                              |=================================                                     |  47%  |                                                                              |===================================                                   |  50%  |                                                                              |=====================================                                 |  53%  |                                                                              |========================================                              |  57%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  63%  |                                                                              |===============================================                       |  67%  |                                                                              |=================================================                     |  70%  |                                                                              |===================================================                   |  73%  |                                                                              |======================================================                |  77%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  83%  |                                                                              |=============================================================         |  87%  |                                                                              |===============================================================       |  90%  |                                                                              |=================================================================     |  93%  |                                                                              |====================================================================  |  97%  |                                                                              |======================================================================| 100%

#>     timestamp provider network_type accuracy latitude longitude altitude
#> 1  1364357009  network         wifi   67.993 43.70667 -72.28910      0.0
#> 2  1364358209  network         wifi   23.000 43.70664 -72.28907      0.0
#> 3  1364359405      gps                16.000 43.70668 -72.28902    136.3
#> 4  1364360609  network         wifi   71.126 43.70667 -72.28909      0.0
#> 5  1364361805      gps                16.000 43.70681 -72.28937    154.5
#> 6  1364363013      gps                12.000 43.70681 -72.28902    130.8
#> 7  1364364209  network         wifi   67.267 43.70668 -72.28909      0.0
#> 8  1364365411      gps                 7.000 43.70681 -72.28937    149.2
#> 9  1364366611      gps                 6.000 43.70678 -72.28935    120.9
#> 10 1364367813      gps                10.000 43.70672 -72.28917    141.1
#> 11 1364369012      gps                12.000 43.70686 -72.28928    151.0
#> 12 1364370209  network         wifi   70.633 43.70668 -72.28909      0.0
#> 13 1364371412      gps                10.000 43.70683 -72.28939    177.0
#> 14 1364372610      gps                 9.000 43.70683 -72.28926    162.2
#> 15 1364373812      gps                10.000 43.70690 -72.28907    116.8
#> 16 1364375012      gps                10.000 43.70669 -72.28916    142.8
#> 17 1364376209  network         wifi   70.688 43.70666 -72.28909      0.0
#> 18 1364377409  network         wifi   67.073 43.70668 -72.28907      0.0
#> 19 1364378612      gps                17.000 43.70682 -72.28907    126.6
#> 20 1364379809  network         wifi  168.684 43.70668 -72.28901      0.0
#> 21 1364381009  network         wifi   84.674 43.70667 -72.28905      0.0
#> 22 1364382209  network         wifi   67.398 43.70667 -72.28909      0.0
#> 23 1364383411      gps                18.000 43.70681 -72.28932    194.9
#> 24 1364384609  network         wifi   20.000 43.70667 -72.28905      0.0
#> 25 1364385811      gps                17.000 43.70697 -72.28922    155.7
#> 26 1364387009  network         wifi   81.328 43.70666 -72.28911      0.0
#> 27 1364387817      gps                15.000 43.70661 -72.28948    154.1
#> 28 1364390609  network         wifi   21.523 43.70455 -72.29478      0.0
#> 29 1364391809  network         wifi   20.911 43.70460 -72.29481      0.0
#> 30 1364393011  network         wifi   36.285 43.70879 -72.28375      0.0
#>    bearing    speed travelstate       date userId source device
#> 1      0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 2      0.0 0.000000      moving 2013-03-27      1    GPS   <NA>
#> 3     96.2 0.250000             2013-03-27      1    GPS   <NA>
#> 4      0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 5      0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 6      0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 7      0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 8      0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 9     96.2 0.559017             2013-03-27      1    GPS   <NA>
#> 10     0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 11     0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 12     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 13     0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 14    96.2 0.750000             2013-03-27      1    GPS   <NA>
#> 15     0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 16     0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 17     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 18     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 19     0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 20     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 21     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 22     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 23     0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 24     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 25     0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 26     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 27     0.0 0.000000             2013-03-27      1    GPS   <NA>
#> 28     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 29     0.0 0.000000  stationary 2013-03-27      1    GPS   <NA>
#> 30     0.0 0.000000      moving 2013-03-27      1    GPS   <NA>
#>    RUNNING_TASKS_baseActivity_mClass RUNNING_TASKS_baseActivity_mPackage
#> 1                               <NA>                                <NA>
#> 2                               <NA>                                <NA>
#> 3                               <NA>                                <NA>
#> 4                               <NA>                                <NA>
#> 5                               <NA>                                <NA>
#> 6                               <NA>                                <NA>
#> 7                               <NA>                                <NA>
#> 8                               <NA>                                <NA>
#> 9                               <NA>                                <NA>
#> 10                              <NA>                                <NA>
#> 11                              <NA>                                <NA>
#> 12                              <NA>                                <NA>
#> 13                              <NA>                                <NA>
#> 14                              <NA>                                <NA>
#> 15                              <NA>                                <NA>
#> 16                              <NA>                                <NA>
#> 17                              <NA>                                <NA>
#> 18                              <NA>                                <NA>
#> 19                              <NA>                                <NA>
#> 20                              <NA>                                <NA>
#> 21                              <NA>                                <NA>
#> 22                              <NA>                                <NA>
#> 23                              <NA>                                <NA>
#> 24                              <NA>                                <NA>
#> 25                              <NA>                                <NA>
#> 26                              <NA>                                <NA>
#> 27                              <NA>                                <NA>
#> 28                              <NA>                                <NA>
#> 29                              <NA>                                <NA>
#> 30                              <NA>                                <NA>
#>    RUNNING_TASKS_id RUNNING_TASKS_numActivities RUNNING_TASKS_numRunning
#> 1                NA                          NA                       NA
#> 2                NA                          NA                       NA
#> 3                NA                          NA                       NA
#> 4                NA                          NA                       NA
#> 5                NA                          NA                       NA
#> 6                NA                          NA                       NA
#> 7                NA                          NA                       NA
#> 8                NA                          NA                       NA
#> 9                NA                          NA                       NA
#> 10               NA                          NA                       NA
#> 11               NA                          NA                       NA
#> 12               NA                          NA                       NA
#> 13               NA                          NA                       NA
#> 14               NA                          NA                       NA
#> 15               NA                          NA                       NA
#> 16               NA                          NA                       NA
#> 17               NA                          NA                       NA
#> 18               NA                          NA                       NA
#> 19               NA                          NA                       NA
#> 20               NA                          NA                       NA
#> 21               NA                          NA                       NA
#> 22               NA                          NA                       NA
#> 23               NA                          NA                       NA
#> 24               NA                          NA                       NA
#> 25               NA                          NA                       NA
#> 26               NA                          NA                       NA
#> 27               NA                          NA                       NA
#> 28               NA                          NA                       NA
#> 29               NA                          NA                       NA
#> 30               NA                          NA                       NA
#>    RUNNING_TASKS_topActivity_mClass RUNNING_TASKS_topActivity_mPackage
#> 1                              <NA>                               <NA>
#> 2                              <NA>                               <NA>
#> 3                              <NA>                               <NA>
#> 4                              <NA>                               <NA>
#> 5                              <NA>                               <NA>
#> 6                              <NA>                               <NA>
#> 7                              <NA>                               <NA>
#> 8                              <NA>                               <NA>
#> 9                              <NA>                               <NA>
#> 10                             <NA>                               <NA>
#> 11                             <NA>                               <NA>
#> 12                             <NA>                               <NA>
#> 13                             <NA>                               <NA>
#> 14                             <NA>                               <NA>
#> 15                             <NA>                               <NA>
#> 16                             <NA>                               <NA>
#> 17                             <NA>                               <NA>
#> 18                             <NA>                               <NA>
#> 19                             <NA>                               <NA>
#> 20                             <NA>                               <NA>
#> 21                             <NA>                               <NA>
#> 22                             <NA>                               <NA>
#> 23                             <NA>                               <NA>
#> 24                             <NA>                               <NA>
#> 25                             <NA>                               <NA>
#> 26                             <NA>                               <NA>
#> 27                             <NA>                               <NA>
#> 28                             <NA>                               <NA>
#> 29                             <NA>                               <NA>
#> 30                             <NA>                               <NA>
#>    mean.accuracy.last30 max.accuracy.last30
#> 1                    NA                  NA
#> 2             67.993000              67.993
#> 3             45.496500              67.993
#> 4             19.500000              23.000
#> 5             36.708667              71.126
#> 6             43.563000              71.126
#> 7             14.000000              16.000
#> 8             39.633500              67.267
#> 9             28.755667              67.267
#> 10             6.500000               7.000
#> 11             8.000000              10.000
#> 12             9.333333              12.000
#> 13            30.877667              70.633
#> 14            30.877667              70.633
#> 15             9.500000              10.000
#> 16             9.500000              10.000
#> 17             9.666667              10.000
#> 18            30.229333              70.688
#> 19            68.880500              70.688
#> 20            42.036500              67.073
#> 21            92.842000             168.684
#> 22            90.119333             168.684
#> 23            76.036000              84.674
#> 24            42.699000              67.398
#> 25            19.000000              20.000
#> 26            18.333333              20.000
#> 27            39.442667              81.328
#> 28            15.000000              15.000
#> 29            21.523000              21.523
#> 30            21.217000              21.523