Adds a variable (or a set of variables) to a dataset by a sliding window approach.

Adds a variable (or a set of variables) to a dataset by a sliding window approach.

slidingWindow(data, utc_col = character(1), fun, steps, time_in_sec,
  check_fun = TRUE, eval_at_rows = numeric(0), unit = "s")

Arguments

data
dataframe.

utc_col
character. Column name of the UTC timestamp.

fun
function. A function that has `data`` as input and a dataframe with 1 row as output.

steps
integer. Number of steps which the sliding window should go back. Use this for data with fixed sample rate.

time_in_sec
integer. Number of seconds which the sliding window should go back. Use this for data with variable sample rate.

check_fun
logical. If TRUE, fun(data) will be evaluated and checked, if the outcome is of correct form. Set to FALSE, if evaluation on the whole dataset takes too long.

eval_at_rows
integer. Integer vector at which rows the function should be executed. Defaults to `1:nrow(data)`.

unit
unit of the timestamps. Can be 's' or 'ms'. Defaults to 's' (seconds).

Value

dataframe with added column(s).

See also

Examples

library(dplyr)

#> 
#> Attache Paket: 'dplyr'

#> The following objects are masked from 'package:lubridate':
#> 
#>     intersect, setdiff, union

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

fun = function(x) data.frame(mean.accuracy.last30 = mean(x$accuracy, na.rm = TRUE),
  max.accuracy.last30 = max(x$accuracy, na.rm = TRUE))
data = studentlife_small %>% filter(!is.na(latitude)) %>% slice(1:30) #needs date_time variable
slidingWindow(data, fun = fun, time_in_sec = 60 * 60,
  utc_col = "timestamp") #mean accuracy of last hour

#>   |                                                                              |                                                                      |   0%  |                                                                              |=====                                                                 |   7%  |                                                                              |=======                                                               |  10%  |                                                                              |=========                                                             |  13%  |                                                                              |============                                                          |  17%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  23%  |                                                                              |===================                                                   |  27%  |                                                                              |=====================                                                 |  30%  |                                                                              |=======================                                               |  33%  |                                                                              |==========================                                            |  37%

#> Warning: kein nicht-fehlendes Argument für max; gebe -Inf zurück

#>   |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  43%  |                                                                              |=================================                                     |  47%  |                                                                              |===================================                                   |  50%  |                                                                              |=====================================                                 |  53%  |                                                                              |========================================                              |  57%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  63%  |                                                                              |===============================================                       |  67%  |                                                                              |=================================================                     |  70%  |                                                                              |===================================================                   |  73%  |                                                                              |======================================================                |  77%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  83%  |                                                                              |=============================================================         |  87%  |                                                                              |===============================================================       |  90%  |                                                                              |=================================================================     |  93%  |                                                                              |====================================================================  |  97%  |                                                                              |======================================================================| 100%

#>    provider network_type accuracy latitude longitude altitude bearing speed
#> 1   network         wifi   24.000 43.75913 -72.32924      0.0       0     0
#> 2       gps                22.000 43.75950 -72.32902    250.9       0     0
#> 3   network         wifi   23.000 43.75913 -72.32924      0.0       0     0
#> 4   network         wifi   25.000 43.75913 -72.32924      0.0       0     0
#> 5   network         wifi   25.000 43.75913 -72.32924      0.0       0     0
#> 6   network         wifi   23.000 43.75913 -72.32924      0.0       0     0
#> 7   network         wifi   25.000 43.75913 -72.32924      0.0       0     0
#> 8   network         cell 5000.000 43.60805 -72.16217      0.0       0     0
#> 9   network         cell 5000.000 43.60805 -72.16217      0.0       0     0
#> 10  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 11  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 12  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 13  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 14  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 15  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 16  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 17  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 18  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 19  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 20  network         cell 3451.000 43.69629 -72.28330      0.0       0     0
#> 21  network         wifi   20.000 43.75913 -72.32923      0.0       0     0
#> 22  network         cell 2294.000 43.70647 -72.28731      0.0       0     0
#> 23  network         wifi   27.443 43.70511 -72.28852      0.0       0     0
#> 24  network         wifi   20.000 43.70672 -72.28730      0.0       0     0
#> 25  network         wifi   20.000 43.70673 -72.28734      0.0       0     0
#> 26  network         wifi   20.000 43.70672 -72.28733      0.0       0     0
#> 27  network         wifi   30.535 43.70672 -72.28729      0.0       0     0
#> 28  network         wifi   20.000 43.70671 -72.28733      0.0       0     0
#> 29  network         wifi   66.924 43.70890 -72.28367      0.0       0     0
#> 30  network         wifi   46.188 43.70895 -72.28397      0.0       0     0
#>    travelstate userId  timestamp device RUNNING_TASKS_baseActivity_mClass
#> 1   stationary     00 1364356963   <NA>                              <NA>
#> 2                  00 1364358144   <NA>                              <NA>
#> 3   stationary     00 1364359364   <NA>                              <NA>
#> 4   stationary     00 1364360563   <NA>                              <NA>
#> 5   stationary     00 1364361763   <NA>                              <NA>
#> 6   stationary     00 1364362963   <NA>                              <NA>
#> 7   stationary     00 1364364163   <NA>                              <NA>
#> 8   stationary     00 1364365364   <NA>                              <NA>
#> 9   stationary     00 1364365368   <NA>                              <NA>
#> 10      moving     00 1364367763   <NA>                              <NA>
#> 11      moving     00 1364367768   <NA>                              <NA>
#> 12      moving     00 1364372563   <NA>                              <NA>
#> 13  stationary     00 1364373763   <NA>                              <NA>
#> 14  stationary     00 1364374963   <NA>                              <NA>
#> 15  stationary     00 1364376163   <NA>                              <NA>
#> 16  stationary     00 1364377363   <NA>                              <NA>
#> 17  stationary     00 1364378563   <NA>                              <NA>
#> 18  stationary     00 1364379763   <NA>                              <NA>
#> 19  stationary     00 1364380963   <NA>                              <NA>
#> 20  stationary     00 1364382163   <NA>                              <NA>
#> 21  stationary     00 1364383363   <NA>                              <NA>
#> 22      moving     00 1364384563   <NA>                              <NA>
#> 23      moving     00 1364385763   <NA>                              <NA>
#> 24      moving     00 1364386963   <NA>                              <NA>
#> 25  stationary     00 1364388163   <NA>                              <NA>
#> 26  stationary     00 1364389363   <NA>                              <NA>
#> 27  stationary     00 1364390563   <NA>                              <NA>
#> 28  stationary     00 1364391763   <NA>                              <NA>
#> 29      moving     00 1364392963   <NA>                              <NA>
#> 30      moving     00 1364394163   <NA>                              <NA>
#>    RUNNING_TASKS_baseActivity_mPackage RUNNING_TASKS_id
#> 1                                 <NA>               NA
#> 2                                 <NA>               NA
#> 3                                 <NA>               NA
#> 4                                 <NA>               NA
#> 5                                 <NA>               NA
#> 6                                 <NA>               NA
#> 7                                 <NA>               NA
#> 8                                 <NA>               NA
#> 9                                 <NA>               NA
#> 10                                <NA>               NA
#> 11                                <NA>               NA
#> 12                                <NA>               NA
#> 13                                <NA>               NA
#> 14                                <NA>               NA
#> 15                                <NA>               NA
#> 16                                <NA>               NA
#> 17                                <NA>               NA
#> 18                                <NA>               NA
#> 19                                <NA>               NA
#> 20                                <NA>               NA
#> 21                                <NA>               NA
#> 22                                <NA>               NA
#> 23                                <NA>               NA
#> 24                                <NA>               NA
#> 25                                <NA>               NA
#> 26                                <NA>               NA
#> 27                                <NA>               NA
#> 28                                <NA>               NA
#> 29                                <NA>               NA
#> 30                                <NA>               NA
#>    RUNNING_TASKS_numActivities RUNNING_TASKS_numRunning
#> 1                           NA                       NA
#> 2                           NA                       NA
#> 3                           NA                       NA
#> 4                           NA                       NA
#> 5                           NA                       NA
#> 6                           NA                       NA
#> 7                           NA                       NA
#> 8                           NA                       NA
#> 9                           NA                       NA
#> 10                          NA                       NA
#> 11                          NA                       NA
#> 12                          NA                       NA
#> 13                          NA                       NA
#> 14                          NA                       NA
#> 15                          NA                       NA
#> 16                          NA                       NA
#> 17                          NA                       NA
#> 18                          NA                       NA
#> 19                          NA                       NA
#> 20                          NA                       NA
#> 21                          NA                       NA
#> 22                          NA                       NA
#> 23                          NA                       NA
#> 24                          NA                       NA
#> 25                          NA                       NA
#> 26                          NA                       NA
#> 27                          NA                       NA
#> 28                          NA                       NA
#> 29                          NA                       NA
#> 30                          NA                       NA
#>    RUNNING_TASKS_topActivity_mClass RUNNING_TASKS_topActivity_mPackage
#> 1                              <NA>                               <NA>
#> 2                              <NA>                               <NA>
#> 3                              <NA>                               <NA>
#> 4                              <NA>                               <NA>
#> 5                              <NA>                               <NA>
#> 6                              <NA>                               <NA>
#> 7                              <NA>                               <NA>
#> 8                              <NA>                               <NA>
#> 9                              <NA>                               <NA>
#> 10                             <NA>                               <NA>
#> 11                             <NA>                               <NA>
#> 12                             <NA>                               <NA>
#> 13                             <NA>                               <NA>
#> 14                             <NA>                               <NA>
#> 15                             <NA>                               <NA>
#> 16                             <NA>                               <NA>
#> 17                             <NA>                               <NA>
#> 18                             <NA>                               <NA>
#> 19                             <NA>                               <NA>
#> 20                             <NA>                               <NA>
#> 21                             <NA>                               <NA>
#> 22                             <NA>                               <NA>
#> 23                             <NA>                               <NA>
#> 24                             <NA>                               <NA>
#> 25                             <NA>                               <NA>
#> 26                             <NA>                               <NA>
#> 27                             <NA>                               <NA>
#> 28                             <NA>                               <NA>
#> 29                             <NA>                               <NA>
#> 30                             <NA>                               <NA>
#>    mean.accuracy.last30 max.accuracy.last30
#> 1                    NA                  NA
#> 2              24.00000              24.000
#> 3              23.00000              24.000
#> 4              22.50000              23.000
#> 5              24.00000              25.000
#> 6              24.33333              25.000
#> 7              24.00000              25.000
#> 8              24.00000              25.000
#> 9            1682.66667            5000.000
#> 10           5000.00000            5000.000
#> 11           4483.66667            5000.000
#> 12                  NaN                -Inf
#> 13           3451.00000            3451.000
#> 14           3451.00000            3451.000
#> 15           3451.00000            3451.000
#> 16           3451.00000            3451.000
#> 17           3451.00000            3451.000
#> 18           3451.00000            3451.000
#> 19           3451.00000            3451.000
#> 20           3451.00000            3451.000
#> 21           3451.00000            3451.000
#> 22           1735.50000            3451.000
#> 23           1157.00000            2294.000
#> 24           1160.72150            2294.000
#> 25             23.72150              27.443
#> 26             20.00000              20.000
#> 27             20.00000              20.000
#> 28             25.26750              30.535
#> 29             25.26750              30.535
#> 30             43.46200              66.924